Add filters to vertical search request (#9)
Adds static filters to the vertical search request.

TEST=manual,unit
J=SLAP-849

Wrote unit tests and included them in this PR. Additionally, created a Typescript file, imported this answers-core library, created a search request with filters, executed a search and printed the response. Compiled and executed the test Typescript code and saw the response as expected.