Add alternative verticals to vertical search response (#17)
When there are no results, an "alternativeVerticals" property may be returned from the API with the vertical results for all verticals in the account with the current query. This PR adds support to shape that into our verticalsearchresponse.

TEST=manual
J=SLAP-849

Conduct a search using a Typescript test page, log console output and see alternativeVerticals come through properly.