Add custom endpoints (#20)
Add custom endpoints to the core

Add a configuration option which allows the user to pass in custom urls which are used for the various services. If an endpoint is not supplied, it falls back to the default. Also connect autocomplete to the top-level of the core library and add an autoComplete function which executes the type of autocomplete request based on the input type.

J=SLAP-886
Test=manual, auto

Use answers-core-testing to supply custom endpoints and confirm that's where the requests are sent. Add automatic tests for universal search and question submission