autocomplete: remove top-level autocomplete function
Removes the top-level autocomplete function from the
autocomplete service.

Since developers who use the core library will already know
about the three different kinds of autocomplete requests,
there is not a need to obfuscate the different kinds of
autocompletes with a top-level function.

It will be very clear what autocomplete function a developer
needs to call with what kind of request based on how functions
are named.

TEST=auto
Ran automated tests and everything passed.