Rename public interfaces (#25)
Rename `Core` and `Config` to `AnswersCore` and `AnswersConfig`

When investigating exported types, I found repos that would attempt to avoid naming conflicts by prefixing interfaces in some way. I think it makes sense for us to rename `Core` and `Config`. If there still is a conflict (or conflicts with other interfaces), developers can rename imports it with the `import module as` syntax.

J=SLAP-897
TEST=manual

Build a site with these new interface names and confirm it works