autocomplete: update autocomplete response domain models
Adds a FilterAutoCompleteResponse to the domain model.
This new model has a sections field that will accommodate
filter autocomplete responses from the backend that contain
a `sections` object.

The presense of a `sections` object is determined by the
`sectioned` field in search paramters in a FilterAutoCompleteRequest.
If `sectioned: true` is passed, then the results from the backend
are grouped by the fields passed in the search paramters.

Updates automated tests to match the new response domain model.

TEST=auto, manual
Ran automated tests and saw them pass.
Used the core testing library with filter autocomplete requests
and saw that the results are parsed correctly.