autocomplete: handle entities in filter response
Adds code to handle when entities are returned as part of a
filter autocomplete response.

Filter autocomplete responses will return entity information
if `fetchEntities` in the search paramters is set to `true`.
In a result object, the entity information will be in a
relatedItem object. We add code to handle and return the
relatedItem object in a result.

Adds an automated test case that tests parsing of a response
with mock entity information.

TEST=auto, manual
Ran automated tests and saw them pass.
Used the core test library to send autocomplete requests and
saw that the responses are what we expect.