Search request documentation (#29)
Documentation for models used when performing a search request

The docs include core, universal search, and vertical search

Currently the interfaces don't link to each other because internally they are default exports, but at the top level of the program they are named exports. A future PR will address this.

I expect that we want to create enums for the SortBys properties in the future, however that change is likely out of scope of this PR

J=SLAP-898
TEST=manual

Inspect the generated docs and make sure they look as expected