Convert all exports to named exports (#33)
Convert all exports to named exports to allow interfaces to properly link to each other in the docs

When the interfaces were exported as default internally but as named exports at the top level of the program, the api-documenter wasn't able to create links between them. This PR fixes that

For the sake of consistency, I also changed all other exports to named exports which has additional benefits:
- It prevents problems involving interfaces with custom default names
- The docs don't specify the export as a default export when it no longer is one

The `Coordinates` interface doesn't have a link to it but that will be fixed in a future PR where when `Coordinates` is renamed

The fixtures are the only default imports because they can't be imported as named imports

J=SLAP-898
TEST=manual

Make sure the tests still pass and that the core library methods all still work. View the docs and confirm the links appear as expected.