Remove the Answers status check from `provideCore`.
We no longer need to include an Answers status check in `provideCore`. This was
confirmed by Product. Because of this, `provideCore` can now be synchronous.
We can also remove the `MasterSwitchApi`.

J=SLAP-976
TEST=manual

Pointed the answers-core-testing repo to my local build of answers-core. I
executed all of the methods exposed in the testing repo to ensure that the
Core functionality still worked as expected.