Ensure that Core can be used within a Node application. (#39)
Previously, Core was not able to be used within a Node app. This was because
we used whatwg-fetch. This library only works within the Browser. It provides
a polyfill of window.fetch. We now use cross-fetch, which provides a platform
agnostic ponyfill of fetch.

Note that the ponyfill is always used. It is not guarded by a check for
window.fetch. Having this check, as we did before, adds complexity to the code.
It does not make any resulting bundle smaller.

I verified that the cross-fetch library did not appreciably increase the size
of the answers-core-testing bundle. So, bundle size should not be noticeably
impacted. As a side benefit, because we're using a ponyfill, Core will not be
polluting the global namespace.

In another PR, I can update the answers-core-testing library to have a Node
application.

J=SLAP-1012
TEST=manual

Verified the following:

- The various tests in the answers-core-testing library still worked as
  expected. This implies that Core still works properly in the Browser.
- Created a Node application that uses all the different Core methods. Things
  worked as expected.