Add ES Modules (#40)
Add support for ES Modules (ES6 Modules)

I followed the common convention of creating two tsconfig.json files which extend a tsconfig.base.json.
The "module" property in the package.json now points to the ES modules and the "main" propery points to the common js modules

J=SLAP-793
TEST=manual

Build the answers-core-testing repo with the es6 modules and confirm that the site works. Run a node script using the commonjs modules and confirm that works.