Make context optional (#42)
Ensure that a vertical request can be sent without requiring context

In order to make the test pass, I updated a few of the transformers to be more resilient to missing response data

J=SLAP-1011
TEST=manual, auto

Test all requests with the answers core library and ensure that none of them require the context param. Add a test for vertical search that verifies that the minimum required params are query and vertical key.