Improve error reporting (#41)
Convert all errors from the api response into an AnswersError object which results in rejected promises

Include basic validation for the response and meta properties of the api response.

This PR removes the AnswersErrors that previously would show up in the question submission or universal search responses.

The ApiResponseValidator consumes ApiResponses and returns either an AnswersError or undefined.

J=SLAP-933
TEST=manual, auto

Use the core testing library to test all the services and test errors such as invalid api keys and missing query inputs and observe the promises reject with the correct error. Check the type and code of the returned error.