Add support for disabled facets (#45)
Support disabled facets by allowing users to supply an empty array of facet options

Update the Facet models for a request and a response models to have a similar structure. Previously they were at two layers of abstraction apart, and this PR puts them at the same level. The facets models also no longer depend on SimpleFilter. The comparator and comparedValue from SimpleFilter move into FacetOption.

The facet response models are prefixed with 'Displayable' and contain extra data meant to be used when displaying facets and fact options on frontends

J=SLAP-1046
TEST=manual, auto

Test sending an empty options array in the testing library and observe the request params. Also observe the response and confirm the shape is correct. Also update the unit tests to account for the new Facet models.