Add IE11 Support (#48)
Add an IE11-compatible UMD bundle which can by imported through @yext/answers-core/legacy

This bundle is compiled to ES5, includes all necessary ponyfills, and includes cross-fetch. It is not minimized per product's request. This bundle only works in the browser because it contains the browser version of cross-fetch which references the DOM.

I restructured the library so that transpiled code is in the lib folder and bundles are in the dist folder. I also modified the api-extractor config to generate a types bundle for just the public types rather than all types.

J=SLAP-1017
TEST=manual

Use the core testing library and import @yext/answers-core/legacy and test the site on chrome and IE11. On IE11, run `const test = new Promise()` and see the message 'Promise is undefined' which indicates that the polyfills are not polluting the global namespace.