use Result model for filtersearch, move SearchParameters to top level (#49)
This commit updates the filterSearch response to use the Result model
instead of it's own special model which has the same information.
Since "source" is not something provided by the api for filtersearch,
it was made into an option parameter for the Result model.

This commit also moves the SearchParameters portion of the
FilterSearchRequest into the top level to reduce nesting.

J=SLAP-975
TEST=manual

tested searching for "", "v", and "virginiaa" using the
answers-core-testing library, and checked that the relatedItem
portions of the core response looked correct