Fix erroring out for allResultsForVertical and alternativeVerticals
The library was erroring out when a vertical search
returned allResultsForVertical or alternativeVerticals.
This was because it expected it was using createVerticalSearchResponse
on these two properties of the response, and by doing so tried to access
a data.meta.uuid property when data.meta is not defined for these two
properties.

TEST=manual, auto

test that I can do a no results search in answers-core-testing