Use files property in package.json instead of npmignore
It's a little simpler to use the files property instead of
npmignore, because of weird interactions npmignore has
with gitignore (i.e. not respecting gitignore).

J=SLAP-840
TEST=manual

tried a local npm install in answers-core-testing, but
it looks like the files property doesn't affect local
npm installs

did a test publish to v1.0.0-beta.6 and npm install of the beta
saw that dist, lib, and legacy folders were the only things
when I npm install-ed the package