Fix facetFilters query param (#65)
Fix the facetFilter query param for vertical searches

In #45 I renamed 'facetFilters' to 'facets', and in doing so I accidentally updated the query param that gets sent to live api which expects 'facetFilters' rather than 'facets'. This meant that facets didn't work, and this PR fixes that.

I also added a test which confirms that all the query params are correct for vertical searches. We had this coverage for universal search but we were missing it for vertical search. In response to some comments, I also cleaned up the tests which ended up being the majority of this PR.

J=none
TEST=auto

Confirm tests pass. Point the SDK cutover branch to this version of the core and observe facets working properly