Target ES5 with the ESM build (#66)
Target ES5 with the ESM build

This PR makes the ESM build compatible with IE11 so long as polyfills are present. Previously, the library would need to be transpiled to ES5 and have polyfills added.

J=SLAP-1074
TEST=manual

Set the core library target to ES5 and include this branch of the core. Remove babel from the build and add `import 'core/js'` to the testing library index. Confirm that the site works on IE11 and Chrome