Add Third Party Notices (#67)
Add THIRD-PARTY-NOTICES file which includes the licenses of our runtime dependencies

This file is generated along with the docs. If the file is out of date, the CI build will fail.

By default, 'generate-license-file' includes the core library license in the file that it generates. I added a script to remove the core library license from the notices file.

J=SLAP-840
TEST=manual

Run npm run build and observe the notices file being generated. Test adding lodash as a dependency and push the package.json without updating the license file and observe the CI build fail.