Add featured snippet (document search) support (#69)
Add support for a new direct answer type: a featured snippet

The direct answer gets a new field: the type field which indicates the type of direct answer. A `DirectAnswerType` enum describes these possible types. I chose to use strings rather than the enums on the models because otherwise the generated docs would create a link to a `DirectAnswerType.FieldValue` or a `DirectAnswerType.FeaturedSnippet` which don't exist in the docs. End users however can use the enums to check the direct answer type.

The `DirectAnswer` models loses the fields specific to the 'FIELD_VALUE' direct answer. Two new DA types are created: `FeaturedSnippetDirectAnswer` and `FieldValueDirectAnswer`. These both extend `DirectAnswer`. A `Snippet` models is created which contains the text and the location of where the direct answer was found.

J=SLAP-1125
TEST=manual, auto

Use the testing library and execute universal searches which contain featured snippet direct answers and field value direct answers. Confirm that the models produced by core match the public interfaces. Add unit tests.