add Coveralls code coverage (#78)
Adds a Coveralls github workflow, also adds a coverage badge
to the readme. I tried looking online for a way to make the coverage
badge reflect the current branch, instead of always master,
but it sounds like there's no way to do that with a static readme.

J=SLAP-873
TEST=auto

we have a coveralls check now (see below) that displays the code coverage percent, wow!
clicking into it links to the coverage for the build, double wow!