add AnswersConfig.additionalQueryParams? property (#79)
This commit adds an additionalQueryParams prop to AnswersConfig.
These additional params will be passed as
query params to every single request made by answers-core.

Because this property is marked as alpha, it is not exported as part of the
AnswersConfig interface, and any typescript users of core will receive
a compilation error if they try to use this property. This property can
still be used in regular js usages of core. In a following PR, we will
add an untrimmed bundle.d.ts which will include alpha and internal
types.

J=SLAP-1320
T=405448
TEST=manual,auto

use local core in the SDK, see that jsLibVersion is passed in search, autocomplete,
and qa submission requests
also try in answers-core-testing