Export the snippet model (#83)
Export the snippet model which was unintentionally missing

When we added FeaturedSnippet direct answers, we accidentally didn't export the Snippet model. This meant that it wasn't showing up in the docs. This PR exports the snippet model.

J=none
TEST=manual

Inspect the updated docs and click through the links to ensure they work