Visitor Support (#109)
Add the visitor option to the AnswersConfig which is included with every request with the exception of question submission requests

J=SLAP-1622
TEST=manual, auto

Use the test site and confirm that the visitor data is being sent correctly to the backend when it's included in the config. See that the question submission endpoint does not include visitor data when fired. Also test that core works as expected when visitor is not included in the config. Add unit tests.