Use api-extractor's type bundle for all builds (#111)
Use api-extractor's bundle.d.ts for all builds of the library

We want to use api-extratcor's bundle.d.ts so that fields marked as alpha, beta, or internal won't be present in output types. Accomplish this by copying the type bundle from dist/bundle.d.ts into lib/esm and lib/commonjs.

Since api-extractor does not yet support bundled declaration maps, I disabled declarationMap in tsconfig.base.json. If we were to leave it on, users of the core library would be incorrectly taken to index.js whenever they click 'Go to Definition' on one of our types or interfaces. This is a confusing DX because index.js does not contain the corresponding type definition. I found that if we disable declarationMap, developers will instead be taken to the correct interface inside of index.d.ts.

J=SLAP-1325
TEST=manual

Use the core test site and confirm that a field marked as alpha (additionalQueryParams from AnswersConfig) does not appear in the autocomplete options when using the AnswersConfig model. Also test clicking the "Go to Definiton" button and see it take me to the corresponding model in index.d.ts