Change how the `Visitor` params are included in API requests.
Originally, the spec called for the `Visitor` object to be passed as a single
param in the Live API requests. Now, the spec has been updated so that each
attribute of the `Visitor` is passed as its own top-level API param.

J=SLAP-1686
TEST=auto

Updated the unit tests and ensured they passed.