Add AnswersConfigWithToken (#114)
After meeting with product, it's decided that breaking change of using types is fine to use. This pr add some tweaks to the config: removing the optional fields in baseAnswersConfig, and some type guard in the services

J=SLAP-1680
TEST=auto

see that it compiled successfully and the updated jest tests passed
have SDK use local core changes, see that request have apiKey in request query string parameters when only apiKey is supplied. see that request have token in authorization header when only token is supplied