Answers-Agent Support (#141)
When hitting any endpoint, pass the Client-SDK header, which includes the ANSWERS_CORE agent specifying the Core version used by the site, as well as any additional agents specified in the request object. If an additional agent has a falsy value for the version, it will not be included in the header. Add unit tests to confirm this behavior.

Because we are importing package.json to get the version, in order to get the `commonjs` and `esm` imports to work correctly, the `outDir` was changed from `lib` to `dist`.

J=SLAP-1947
TEST=auto, manual

Check that Jest tests pass and see that the correct header is passed when using the test-site.