Upgrade Jest to v28 (#167)
This PR upgrades Jest from v26 to v28. In v27, there was a breaking change that required the `testEnvironment` to be manually set to `jsdom` and in v28, to use that environment, `jest-environment-jsdom` now has to be installed separately. `webpack` and `ts-loader` versions were also updated in response to webpack errors when making the legacy bundle.

J=SLAP-2123
TEST=auto

Check that there are no errors during the build and that the Jest tests still pass.