Add matchedSubstrings to spellCheck interface (#192)
Raiden needs access](https://yext.slack.com/archives/C03NQ9X8ACB/p1658505872261039) to `matchedSubstrings` from the spellCheck part of the Search API response.  Since changing `correctedQuery` from a string to an object would be a breaking change, we added a new property to the spellCheck interface that holds the `matchedSubstrings`.  `matchedSubstrings` was also added to the Core response when the spellCheck object is created in `createSpellCheck`.

J=SLAP-2274
TEST=manual

It was verified that when spellCheck was provided in the Core response, that `matchedSubstrings` was also provided as its own property.