Add `StringValueDirectAnswer` interface.
This PR creates the `StringValueDirectAnswer` interface, which
describes a direct answer sourced from a string (or string list) field
in the Knowledge Graph.

J=SLAP-2315
TEST=manual

I made a few code changes locally which are not included in this PR. These
changes allowed me to generate a switch statement on `fieldType`. I added
cases for `single_line_text` and `multi_line_text`. These cases were hit as
expected when getting direct answers for string and string list fields. In
these switch statements, I saw the IDE correctly inferring that the DA was a
`StringValueDirectAnswer`.