Update static filters (#204)
Update the `StaticFilter` interface to only allow combinations of filters supported by the backend. After talking with Watson, only `AND`s of `OR`s (conjunctions of disjunctions), or structures that can be trivially converted to such, are accepted. To reflect this, `ConjunctionStaticFilter` and `DisjunctionStaticFilter` models were created, with the latter having limitations on which filters it can combine so a disjunction of conjunctions is not allowed. Also, I added some Jest tests for a couple more edge cases that were being missed when serializing a static filter.

J=SLAP-2325
TEST=auto, manual

See that the new and existing Jest tests pass. Spin up the test-site and successfully fire vertical searches with static filters defined with the new interface.