Allow Webpack & Rollup to tree-shake Core.
This PR adds `"sideEffects": false` to the package.json. This
flag lets Webpack & Rollup know that search-core is safe to tree-shake.
Prior to this change, a PagesJS user was seeing the entirety of Core
included in their bundles, even though they were only using a small part
of its exported functionality.