Add optional name property to near field value filters
During filter search, the value of a near filter in a result corresponds to the name of the location for that filter. We already support setting the "name" property on a near filter in the backend, so we can start setting the "name" property on the frontend, too.

Tested by changing the filterSearchRequest in test-site's autocompleteRequests.ts to query the Locations vertical and builtin.location field on location entities using the input "vir". Saw that the name field was set on the resulting near field value filters to whatever the location of the filter was (e.g. "Virginia, USA").