Support facetAllowlist query parameter for vertical searches (#310)
Allows the "facetAllowlist" parameter to be set on vertical searches. This parameter specifies the subset of field IDs that are configured as facet fields that facet options should be returned for in the search response. If unspecified, all fields configured as facetable can be returned if they have data for the given query.

J=WAT-4928
TEST=manual

Added a query to the test-site vertical requests that returns facets. Confirmed that the facetAllowlist parameter is settable and works as expected.